Isolate real-worker acceptance tests
Changed:
- Move real-worker overlay flows into a clean pytest process and update the test catalogue.
- Give idle-renewal and mocked provider tests independent scheduling margins.

Fixed:
- Kill timed-out Windows pytest process trees with the native tree terminator.
- Prevent macOS scheduling delay from racing the unrelated stream hard-cap assertion.

Notes:
- Product runtime behavior and translations are unchanged.
