+++ tests/TEST_MAP.md
@@ -4,7 +4,7 @@ Generated by `python scripts/test_map.py`. Do not edit this report by hand.
 
 Every pytest test receives exactly one behaviour marker (`internal` or `user_path`) and exactly one execution marker (`github_safe` or `isolated_host`) from `tests/catalog/test_map.json`.
 
-- Pytest files: **169**
+- Pytest files: **170**
 - Collected pytest nodes: **2666**
 - Manual diagnostic files: **3**
 - Unclassified files: **0**
@@ -15,7 +15,7 @@ Every pytest test receives exactly one behaviour marker (`internal` or `user_pat
 | Category | Meaning | Files | Test nodes |
 |---|---|---:|---:|
 | GI | GitHub-safe internal | 137 | 1897 |
-| GU | GitHub-safe user path | 29 | 756 |
+| GU | GitHub-safe user path | 30 | 756 |
 | II | Isolated-host internal | 5 | 4 |
 | IU | Isolated-host user path | 1 | 9 |
 
@@ -27,7 +27,7 @@ Every pytest test receives exactly one behaviour marker (`internal` or `user_pat
 | component | 31 | 341 |
 | integration | 61 | 1195 |
 | ui | 15 | 390 |
-| e2e | 14 | 366 |
+| e2e | 15 | 366 |
 | native | 19 | 167 |
 | support | 15 | 132 |
 | catalog | 5 | 14 |
@@ -160,7 +160,8 @@ Every pytest test receives exactly one behaviour marker (`internal` or `user_pat
 | `tests/test_optional_install_acceptance.py` | pytest | e2e | GU | windows, linux, macos | github-actions | 3 |
 | `tests/test_optional_install_dialog.py` | pytest | ui | GU | windows, linux, macos | github-actions | 23 |
 | `tests/test_overlay_bubble_visibility.py` | pytest | ui | GU | windows, linux, macos | github-actions | 11 |
-| `tests/test_overlay_shell_acceptance.py` | pytest | e2e | GU | windows, linux, macos | github-actions | 6 |
+| `tests/test_overlay_shell_acceptance.py` | pytest | e2e | GU | windows, linux, macos | github-actions | 2 |
+| `tests/test_overlay_worker_acceptance.py` | pytest | e2e | GU | windows, linux, macos | github-actions | 4 |
 | `tests/test_page_context_extraction.py` | pytest | unit | GI | windows, linux, macos | github-actions | 6 |
 | `tests/test_planned_chunking.py` | pytest | component | GI | windows, linux, macos | github-actions | 4 |
 | `tests/test_platform_macos.py` | pytest | native | GI | macos | github-actions | 19 |
