+++ tests/catalog/test_map.json
@@ -1874,7 +1874,22 @@
         "macos"
       ],
       "schedule": "github-actions",
-      "node_count": 6
+      "node_count": 2
+    },
+    {
+      "path": "tests/test_overlay_worker_acceptance.py",
+      "kind": "pytest",
+      "scope": "e2e",
+      "behaviour": "user_path",
+      "execution": "github_safe",
+      "category": "GU",
+      "platforms": [
+        "windows",
+        "linux",
+        "macos"
+      ],
+      "schedule": "github-actions",
+      "node_count": 4
     },
     {
       "path": "tests/test_page_context_extraction.py",
