+++ tests/workflows/feature_acceptance.json
@@ -383,7 +383,7 @@
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Last chat tray QAction -> show_last_chat signal -> QtProtocolHost._show_chat",
       "test_node_ids": [
-        "tests/test_overlay_shell_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"
+        "tests/test_overlay_worker_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"
       ],
       "success_assertions": [
         "The real Last chat action makes the production ChatWindow visible in a four-worker runtime."
@@ -400,7 +400,7 @@
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Memory tray QAction -> supervisor FlowController.open_memory -> MemoryViewer",
       "test_node_ids": [
-        "tests/test_overlay_shell_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"
+        "tests/test_overlay_worker_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"
       ],
       "success_assertions": [
         "The Memory QAction crosses UI event IPC and the real brain memory-list call before the MemoryViewer becomes visible."
@@ -417,7 +417,7 @@
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Addon Manager tray QAction -> supervisor FlowController.open_addons -> add-on dialog",
       "test_node_ids": [
-        "tests/test_overlay_shell_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"
+        "tests/test_overlay_worker_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"
       ],
       "success_assertions": [
         "The Addon Manager QAction crosses UI event IPC and the real brain add-on-list call before the manager becomes visible."
@@ -434,7 +434,7 @@
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Settings tray QAction -> supervisor FlowController.open_settings -> SettingsDialog",
       "test_node_ids": [
-        "tests/test_overlay_shell_acceptance.py::test_real_worker_settings_action_finishes_loading_then_quit"
+        "tests/test_overlay_worker_acceptance.py::test_real_worker_settings_action_finishes_loading_then_quit"
       ],
       "success_assertions": [
         "The real Settings QAction crosses UI event IPC and the production SettingsDialog finishes its deferred page build and value-loading pass."
@@ -451,7 +451,7 @@
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "visible provider badge QPushButton -> IconOverlay._open_harness_controls -> HarnessControlsDialog",
       "test_node_ids": [
-        "tests/test_overlay_shell_acceptance.py::test_real_worker_provider_action_opens_matching_controls_then_quit"
+        "tests/test_overlay_worker_acceptance.py::test_real_worker_provider_action_opens_matching_controls_then_quit"
       ],
       "success_assertions": [
         "Clicking the real provider badge opens ChatGPT controls in Codex mode and Claude controls in Claude mode."
@@ -468,7 +468,7 @@
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Runtime Status tray QAction -> supervisor snapshot -> live runtime diagnostics dialog",
       "test_node_ids": [
-        "tests/test_overlay_shell_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"
+        "tests/test_overlay_worker_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"
       ],
       "success_assertions": [
         "The Runtime Status QAction crosses supervisor IPC and displays the live diagnostics window while all workers are active."
@@ -485,7 +485,7 @@
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Quit tray QAction -> QApplication.quit -> UI worker exit",
       "test_node_ids": [
-        "tests/test_overlay_shell_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"
+        "tests/test_overlay_worker_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"
       ],
       "success_assertions": [
         "Triggering the real Quit QAction exits the actual UI worker, after which supervisor shutdown leaves no remaining worker."
@@ -3584,7 +3584,7 @@
       "production_entry_point": "visible floating ChatGPT/Claude provider badge -> deferred HarnessControlsDialog",
       "test_node_ids": [
         "tests/test_provider_controls_acceptance.py::test_floating_provider_badge_opens_and_saves_real_controls",
-        "tests/test_overlay_shell_acceptance.py::test_real_worker_provider_action_opens_matching_controls_then_quit"
+        "tests/test_overlay_worker_acceptance.py::test_real_worker_provider_action_opens_matching_controls_then_quit"
       ],
       "success_assertions": [
         "Each visible provider badge opens the matching real controls window; its Save button persists, reloads, closes, and emits the live settings payload."
