+++ tests/workflows/feature_acceptance_overrides.json
@@ -211,7 +211,7 @@
       "function_id": "F022",
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Last chat tray QAction -> show_last_chat signal -> QtProtocolHost._show_chat",
-      "test_node_ids": ["tests/test_overlay_shell_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"],
+      "test_node_ids": ["tests/test_overlay_worker_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"],
       "success_assertions": ["The real Last chat action makes the production ChatWindow visible in a four-worker runtime."],
       "audit_note": "Runs through a real isolated IPC worker tree and waits for this surface before triggering the next action.",
       "dependency_status": "audited"
@@ -220,7 +220,7 @@
       "function_id": "F023",
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Memory tray QAction -> supervisor FlowController.open_memory -> MemoryViewer",
-      "test_node_ids": ["tests/test_overlay_shell_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"],
+      "test_node_ids": ["tests/test_overlay_worker_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"],
       "success_assertions": ["The Memory QAction crosses UI event IPC and the real brain memory-list call before the MemoryViewer becomes visible."],
       "audit_note": "Uses real UI and brain workers, not a signal recorder.",
       "dependency_status": "audited"
@@ -229,7 +229,7 @@
       "function_id": "F024",
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Addon Manager tray QAction -> supervisor FlowController.open_addons -> add-on dialog",
-      "test_node_ids": ["tests/test_overlay_shell_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"],
+      "test_node_ids": ["tests/test_overlay_worker_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"],
       "success_assertions": ["The Addon Manager QAction crosses UI event IPC and the real brain add-on-list call before the manager becomes visible."],
       "audit_note": "Uses real UI and brain workers, not the singleton helper alone.",
       "dependency_status": "audited"
@@ -238,7 +238,7 @@
       "function_id": "F025",
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Settings tray QAction -> supervisor FlowController.open_settings -> SettingsDialog",
-      "test_node_ids": ["tests/test_overlay_shell_acceptance.py::test_real_worker_settings_action_finishes_loading_then_quit"],
+      "test_node_ids": ["tests/test_overlay_worker_acceptance.py::test_real_worker_settings_action_finishes_loading_then_quit"],
       "success_assertions": ["The real Settings QAction crosses UI event IPC and the production SettingsDialog finishes its deferred page build and value-loading pass."],
       "audit_note": "Settings runs alone in a fresh worker tree and must report ready rather than merely visible.",
       "dependency_status": "audited"
@@ -247,7 +247,7 @@
       "function_id": "F026",
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "visible provider badge QPushButton -> IconOverlay._open_harness_controls -> HarnessControlsDialog",
-      "test_node_ids": ["tests/test_overlay_shell_acceptance.py::test_real_worker_provider_action_opens_matching_controls_then_quit"],
+      "test_node_ids": ["tests/test_overlay_worker_acceptance.py::test_real_worker_provider_action_opens_matching_controls_then_quit"],
       "success_assertions": ["Clicking the real provider badge opens ChatGPT controls in Codex mode and Claude controls in Claude mode."],
       "audit_note": "Both behaviorally distinct provider badge modes run in separate real four-worker runtimes.",
       "dependency_status": "audited",
@@ -257,7 +257,7 @@
       "function_id": "F027",
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Runtime Status tray QAction -> supervisor snapshot -> live runtime diagnostics dialog",
-      "test_node_ids": ["tests/test_overlay_shell_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"],
+      "test_node_ids": ["tests/test_overlay_worker_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"],
       "success_assertions": ["The Runtime Status QAction crosses supervisor IPC and displays the live diagnostics window while all workers are active."],
       "audit_note": "This is the supported pure-Python UI-host path.",
       "dependency_status": "audited"
@@ -266,7 +266,7 @@
       "function_id": "F028",
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "real UI-worker Quit tray QAction -> QApplication.quit -> UI worker exit",
-      "test_node_ids": ["tests/test_overlay_shell_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"],
+      "test_node_ids": ["tests/test_overlay_worker_acceptance.py::test_real_worker_tray_actions_open_each_shared_surface_then_quit"],
       "success_assertions": ["Triggering the real Quit QAction exits the actual UI worker, after which supervisor shutdown leaves no remaining worker."],
       "audit_note": "The test observes the process exit rather than a mocked quit callback.",
       "dependency_status": "audited"
@@ -2000,7 +2000,7 @@
       "function_id": "F196",
       "acceptance_status": "real_entry_accepted",
       "production_entry_point": "visible floating ChatGPT/Claude provider badge -> deferred HarnessControlsDialog",
-      "test_node_ids": ["tests/test_provider_controls_acceptance.py::test_floating_provider_badge_opens_and_saves_real_controls", "tests/test_overlay_shell_acceptance.py::test_real_worker_provider_action_opens_matching_controls_then_quit"],
+      "test_node_ids": ["tests/test_provider_controls_acceptance.py::test_floating_provider_badge_opens_and_saves_real_controls", "tests/test_overlay_worker_acceptance.py::test_real_worker_provider_action_opens_matching_controls_then_quit"],
       "success_assertions": ["Each visible provider badge opens the matching real controls window; its Save button persists, reloads, closes, and emits the live settings payload."],
       "audit_note": "The in-process case proves Save/reload and the process-level UI-worker case proves the badge RPC reaches the production window.",
       "dependency_status": "audited",
